backend/requirements.txt
action-msgs==1.2.1
actionlib-msgs==4.2.4
actuator-msgs==0.0.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
ament-cmake-test==1.3.11
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
builtin-interfaces==1.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
composition-interfaces==1.2.1
diagnostic-msgs==4.2.4
Django==5.2.8
django-cors-headers==4.9.0
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
domain-coordinator==0.10.0
frozenlist==1.8.0
geometry-msgs==4.2.4
gps-msgs==2.0.4
idna==3.11
interactive-markers==2.3.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
laser-geometry==2.4.0
launch==1.0.7
launch-ros==0.19.8
launch-testing==1.0.7
launch-testing-ros==0.19.8
launch-xml==1.0.7
launch-yaml==1.0.7
lifecycle-msgs==1.2.1
map-msgs==2.1.0
message-filters==4.3.5
multidict==6.7.0
nav-msgs==4.2.4
osrf-pycommon==2.1.4
phonenumbers==9.0.18
pillow==12.0.0
propcache==0.4.1
PyJWT==2.10.1
python-decouple==3.8
python-qt-binding==1.1.2
rcl-interfaces==1.2.1
rclpy==3.3.15
rcutils==5.1.6
requests==2.32.5
resource-retriever==3.1.2
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.16
ros2action==0.18.11
ros2bag==0.15.13
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.8
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.13
rosbag2-py==0.15.13
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
sqlparse==0.5.3
sros2==0.10.5
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
tf2-geometry-msgs==0.25.9
tf2-kdl==0.25.9
tf2-msgs==0.25.9
tf2-py==0.25.9
tf2-ros-py==0.25.9
tf2-tools==0.25.9
trajectory-msgs==4.2.4
twilio==9.8.5
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
urllib3==2.5.0
vision-msgs==4.1.1
visualization-msgs==4.2.4
xacro==2.0.8
yarl==1.22.0
